BoardViewのAPIをプロトコルに切り出して、Reversi logicsをそちらに実装
ただし、placeDisk、お前はだめだ（あとで考える）
