+++ Reversi/Board.swift
@@ -1,0 +1,137 @@
+import Foundation
+
+protocol Board {
+    /// 盤の幅（ `8` ）を表します。
+    var width: Int { get }
+    
+    /// 盤の高さ（ `8` ）を返します。
+    var height: Int { get }
+    
+    /// 盤のセルの `x` の範囲（ `0 ..< 8` ）を返します。
+    var xRange: Range<Int> { get }
+    
+    /// 盤のセルの `y` の範囲（ `0 ..< 8` ）を返します。
+    var yRange: Range<Int> { get }
+
+    /// 盤をゲーム開始時に状態に戻します。このメソッドはアニメーションを伴いません。
+    func reset()
+    
+    /// `x`, `y` で指定されたセルの状態を返します。
+    /// セルにディスクが置かれていない場合、 `nil` が返されます。
+    /// - Parameter x: セルの列です。
+    /// - Parameter y: セルの行です。
+    /// - Returns: セルにディスクが置かれている場合はそのディスクの値を、置かれていない場合は `nil` を返します。
+    func diskAt(x: Int, y: Int) -> Disk?
+
+    /// `x`, `y` で指定されたセルの状態を、与えられた `disk` に変更します。
+    /// `animated` が `true` の場合、アニメーションが実行されます。
+    /// アニメーションの完了通知は `completion` ハンドラーで受け取ることができます。
+    /// - Parameter disk: セルに設定される新しい状態です。 `nil` はディスクが置かれていない状態を表します。
+    /// - Parameter x: セルの列です。
+    /// - Parameter y: セルの行です。
+    /// - Parameter animated: セルの状態変更を表すアニメーションを表示するかどうかを指定します。
+    /// - Parameter completion: アニメーションの完了通知を受け取るハンドラーです。
+    ///     `animated` に `false` が指定された場合は状態が変更された後で即座に同期的に呼び出されます。
+    ///     ハンドラーが受け取る `Bool` 値は、 `UIView.animate()`  等に準じます。
+    func setDisk(_ disk: Disk?, atX x: Int, y: Int, animated: Bool, completion: ((Bool) -> Void)?)
+}
+
+extension Board {
+    /// `side` で指定された色のディスクが盤上に置かれている枚数を返します。
+    /// - Parameter side: 数えるディスクの色です。
+    /// - Returns: `side` で指定された色のディスクの、盤上の枚数です。
+    func countDisks(of side: Disk) -> Int {
+        var count = 0
+        
+        for y in yRange {
+            for x in xRange {
+                if diskAt(x: x, y: y) == side {
+                    count +=  1
+                }
+            }
+        }
+        
+        return count
+    }
+    
+    /// 盤上に置かれたディスクの枚数が多い方の色を返します。
+    /// 引き分けの場合は `nil` が返されます。
+    /// - Returns: 盤上に置かれたディスクの枚数が多い方の色です。引き分けの場合は `nil` を返します。
+    func sideWithMoreDisks() -> Disk? {
+        let darkCount = countDisks(of: .dark)
+        let lightCount = countDisks(of: .light)
+        if darkCount == lightCount {
+            return nil
+        } else {
+            return darkCount > lightCount ? .dark : .light
+        }
+    }
+    
+    // 一時的に private 外しておく。あとで考える。
+    /* private */ func flippedDiskCoordinatesByPlacingDisk(_ disk: Disk, atX x: Int, y: Int) -> [(Int, Int)] {
+        let directions = [
+            (x: -1, y: -1),
+            (x:  0, y: -1),
+            (x:  1, y: -1),
+            (x:  1, y:  0),
+            (x:  1, y:  1),
+            (x:  0, y:  1),
+            (x: -1, y:  0),
+            (x: -1, y:  1),
+        ]
+        
+        guard diskAt(x: x, y: y) == nil else {
+            return []
+        }
+        
+        var diskCoordinates: [(Int, Int)] = []
+        
+        for direction in directions {
+            var x = x
+            var y = y
+            
+            var diskCoordinatesInLine: [(Int, Int)] = []
+            flipping: while true {
+                x += direction.x
+                y += direction.y
+                
+                switch (disk, diskAt(x: x, y: y)) { // Uses tuples to make patterns exhaustive
+                case (.dark, .some(.dark)), (.light, .some(.light)):
+                    diskCoordinates.append(contentsOf: diskCoordinatesInLine)
+                    break flipping
+                case (.dark, .some(.light)), (.light, .some(.dark)):
+                    diskCoordinatesInLine.append((x, y))
+                case (_, .none):
+                    break flipping
+                }
+            }
+        }
+        
+        return diskCoordinates
+    }
+    
+    /// `x`, `y` で指定されたセルに、 `disk` が置けるかを調べます。
+    /// ディスクを置くためには、少なくとも 1 枚のディスクをひっくり返せる必要があります。
+    /// - Parameter x: セルの列です。
+    /// - Parameter y: セルの行です。
+    /// - Returns: 指定されたセルに `disk` を置ける場合は `true` を、置けない場合は `false` を返します。
+    func canPlaceDisk(_ disk: Disk, atX x: Int, y: Int) -> Bool {
+        !flippedDiskCoordinatesByPlacingDisk(disk, atX: x, y: y).isEmpty
+    }
+    
+    /// `side` で指定された色のディスクを置ける盤上のセルの座標をすべて返します。
+    /// - Returns: `side` で指定された色のディスクを置ける盤上のすべてのセルの座標の配列です。
+    func validMoves(for side: Disk) -> [(x: Int, y: Int)] {
+        var coordinates: [(Int, Int)] = []
+        
+        for y in yRange {
+            for x in xRange {
+                if canPlaceDisk(side, atX: x, y: y) {
+                    coordinates.append((x, y))
+                }
+            }
+        }
+        
+        return coordinates
+    }
+}
